requirements.txt
streamlit
pandas
psutil
apcaccess
jsonlines
netifaces
dacite
python-dateutil
plotly